requirements.txt
flask
line-bot-sdk
google-generativeai
psycopg2-binary
sqlalchemy
pgvector
python-dotenv
gunicorn
Pillow
requests
beautifulsoup4
langchain-text-splitters